wip: isolate conversation routing fallback
